refactor(#343): 리프레시 토큰 재발급 로직 refreshAndStoreToken 함수로 분리
