+++ src/api/auth.ts
@@ -1,13 +1,9 @@
 import axios from "axios";
 
-
-
-import { postAuthRefresh } from "@/@generated/api";
 import ko from "@/constants/ko.json";
-import { STORAGE_KEY } from "@/constants/storage";
 import { notify } from "@/hooks/notify";
 import { connectSocket } from "@/socket";
-import { clearTokens, getAccessToken, getRefreshToken } from "@/utils/storage";
+import { clearTokens, getAccessToken } from "@/utils/storage";
 
 
 
@@ -16,6 +12,7 @@ import { BASE_URL } from "./api";
 
 
 import type { AxiosInstance, InternalAxiosRequestConfig } from "axios";
+import { refreshAndStoreToken } from "@/utils";
 
 
 /* AuthAxios 인스턴스 생성 */
@@ -57,29 +54,15 @@ AuthAxios.interceptors.response.use(
 
       try {
         /* 토큰 재발급 요청 */
-        const refreshToken = getRefreshToken();
-        if (!refreshToken) throw new Error("리프레시 토큰이 없습니다.");
-
-        const response = await postAuthRefresh({ refreshToken });
-        
-        if (!response.data) throw new Error("토큰 재발급 응답에 데이터가 없습니다.");
-
-        const {
-            accessToken: newAccessToken,
-            refreshToken: newRefreshToken,
-        } = response.data;
+        const newAccessToken = await refreshAndStoreToken();
+        if (!newAccessToken)
+          throw new Error("재발급된 accessToken이 없습니다.");
 
         const originRequest = config; // 이전 요청 저장
-
-        // 로컬 또는 세션에 재발급된 토큰 저장
-        const storage = localStorage.getItem(STORAGE_KEY.accessToken)
-          ? localStorage
-          : sessionStorage;
-        storage.setItem(STORAGE_KEY.accessToken, newAccessToken ?? "");
-        storage.setItem(STORAGE_KEY.refreshToken, newRefreshToken ?? "");
-
-        connectSocket();
         originRequest.headers["Authorization"] = `Bearer ${newAccessToken}`;
+        
+        connectSocket();
+
         return axios(originRequest);
       } catch (reissueError: any) {
         if (reissueError.response && reissueError.response.status === 404) {
